prepare for enabling fork tests
