+++ core/tests/blockchain_repair_test.go
@@ -29,7 +29,6 @@ import (
 
 	"github.com/golang/mock/gomock"
 
-	"github.com/ethereum/go-ethereum/accounts"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/consensus/bor"
 	"github.com/ethereum/go-ethereum/consensus/bor/api"
@@ -1788,6 +1787,8 @@ func testRepair(t *testing.T, tt *rewindTest, snapshots bool) {
 	}
 	defer db.Close() // Might double close, should be fine
 
+	chainConfig := params.BorUnittestChainConfig
+
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
@@ -1798,7 +1799,7 @@ func testRepair(t *testing.T, tt *rewindTest, snapshots bool) {
 	spanner.EXPECT().GetCurrentValidators(gomock.Any(), gomock.Any()).Return([]*valset.Validator{
 		{
 			ID:               0,
-			Address:          common.Address{0x1},
+			Address:          miner.TestBankAddress,
 			VotingPower:      100,
 			ProposerPriority: 0,
 		},
@@ -1809,44 +1810,15 @@ func testRepair(t *testing.T, tt *rewindTest, snapshots bool) {
 
 	contractMock := bor.NewMockGenesisContract(ctrl)
 
-	// Initialize a fresh chain
-	var (
-		gspec = &core.Genesis{
-			Config:  params.BorUnittestChainConfig,
-			BaseFee: big.NewInt(params.InitialBaseFee),
-		}
-		config = &core.CacheConfig{
-			TrieCleanLimit: 256,
-			TrieDirtyLimit: 256,
-			TrieTimeLimit:  5 * time.Minute,
-			SnapshotLimit:  0, // Disable snapshot by default
-		}
-	)
-
-	engine := miner.NewFakeBor(t, db, params.BorUnittestChainConfig, ethAPIMock, spanner, heimdallClientMock, contractMock)
+	engine := miner.NewFakeBor(t, db, chainConfig, ethAPIMock, spanner, heimdallClientMock, contractMock)
 	defer engine.Close()
 
-	engineBorInternal, ok := engine.(*bor.Bor)
-	if ok {
-		gspec.ExtraData = make([]byte, 32+common.AddressLength+crypto.SignatureLength)
-		copy(gspec.ExtraData[32:32+common.AddressLength], testAddress1.Bytes())
+	chainConfig.LondonBlock = big.NewInt(0)
 
-		engineBorInternal.Authorize(testAddress1, func(account accounts.Account, s string, data []byte) ([]byte, error) {
-			return crypto.Sign(crypto.Keccak256(data), testKey1)
-		})
-	}
+	_, back, closeFn := miner.NewTestWorker(t, chainConfig, engine, db, 0)
+	defer closeFn()
 
-	genesis := gspec.MustCommit(db)
-
-	if snapshots {
-		config.SnapshotLimit = 256
-		config.SnapshotWait = true
-	}
-
-	chain, err := core.NewBlockChain(db, config, params.BorUnittestChainConfig, engine, vm.Config{}, nil, nil)
-	if err != nil {
-		t.Fatalf("Failed to create chain: %v", err)
-	}
+	genesis := back.BlockChain().Genesis()
 
 	// If sidechain blocks are needed, make a light chain and import it
 	var sideblocks types.Blocks
@@ -1855,42 +1827,42 @@ func testRepair(t *testing.T, tt *rewindTest, snapshots bool) {
 			b.SetCoinbase(testAddress1)
 
 			if bor.IsSprintStart(b.Number().Uint64(), params.BorUnittestChainConfig.Bor.Sprint) {
-				b.SetExtra(gspec.ExtraData)
+				b.SetExtra(back.Genesis.ExtraData)
 			} else {
 				b.SetExtra(make([]byte, 32+crypto.SignatureLength))
 			}
 		})
-		if _, err := chain.InsertChain(sideblocks); err != nil {
+		if _, err := back.BlockChain().InsertChain(sideblocks); err != nil {
 			t.Fatalf("Failed to import side chain: %v", err)
 		}
 	}
 
 	canonblocks, _ := core.GenerateChain(params.BorUnittestChainConfig, genesis, engine, rawdb.NewMemoryDatabase(), tt.canonicalBlocks, func(i int, b *core.BlockGen) {
-		b.SetCoinbase(common.Address{0x02})
+		b.SetCoinbase(miner.TestBankAddress)
 		b.SetDifficulty(big.NewInt(1000000))
 
 		if bor.IsSprintStart(b.Number().Uint64(), params.BorUnittestChainConfig.Bor.Sprint) {
-			b.SetExtra(gspec.ExtraData)
+			b.SetExtra(back.Genesis.ExtraData)
 		} else {
 			b.SetExtra(make([]byte, 32+crypto.SignatureLength))
 		}
 	})
-	if _, err := chain.InsertChain(canonblocks[:tt.commitBlock]); err != nil {
+	if _, err := back.BlockChain().InsertChain(canonblocks[:tt.commitBlock]); err != nil {
 		t.Fatalf("Failed to import canonical chain start: %v", err)
 	}
 	if tt.commitBlock > 0 {
-		err = chain.StateCache().TrieDB().Commit(canonblocks[tt.commitBlock-1].Root(), true, nil)
+		err = back.BlockChain().StateCache().TrieDB().Commit(canonblocks[tt.commitBlock-1].Root(), true, nil)
 		if err != nil {
 			t.Fatal("on trieDB.Commit", err)
 		}
 
 		if snapshots {
-			if err := chain.Snaps().Cap(canonblocks[tt.commitBlock-1].Root(), 0); err != nil {
+			if err := back.BlockChain().Snaps().Cap(canonblocks[tt.commitBlock-1].Root(), 0); err != nil {
 				t.Fatalf("Failed to flatten snapshots: %v", err)
 			}
 		}
 	}
-	if _, err := chain.InsertChain(canonblocks[tt.commitBlock:]); err != nil {
+	if _, err := back.BlockChain().InsertChain(canonblocks[tt.commitBlock:]); err != nil {
 		t.Fatalf("Failed to import canonical chain tail: %v", err)
 	}
 	// Force run a freeze cycle
